# storage/api-docs/api-docs-backup.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Pague Fácil",
        "version": "1.0.1"
    },
    "schemes": [
        "https",
        "http"
    ],
    "paths": {
        "/api/auth/customer/register": {
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Registro público de clientes",
                "description": "Retorna a coleção de dados do cliente registrado",
                "operationId": "bcc02973c50b2eb03d8efe1b2055c6c7",
                "requestBody": {
                "description": "Cliente objeto que precisa ser adicionado",
                "content": {
                    "application/json": {
                      "schema": {
                        "$ref": "#/definitions/customer"
                      }
                    },
                    "required": true
                }
                },

                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/api/auth/customer": {
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Autênticação do cliente",
                "description": "Retorna o token de acesso",
                "parameters": [
                    {
                      "name": "email",
                      "in": "query",
                      "description": "The E-mail for login",
                      "required": true,
                      "type": "string"
                    },
                    {
                      "name": "password",
                      "in": "query",
                      "description": "The password for login in clear text",
                      "required": true,
                      "type": "string"
                    }
                ],
                "operationId": "5a0ecf4a2aa28bb591fc3713c0a6beb4",
                "responses": {
                    "200": {
                        "description": "clientes"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "customer": {
          "type": "object",
          "properties": {
            "type": {
              "type": "integer",
              "format": "int64",
              "example": "1"
            },
            "cpf": {
                "type": "integer",
                "format": "int64",
                "example": "12345678910"
              },
            "first_name": {
              "type": "string",
              "example": "Jonh"
            },
            "last_name": {
                "type": "string",
                "example": "Doe"
            },
            "email": {
                "type": "string",
                "example": "joedoe@example.com"
            },
            "password": {
                "type": "string",
                "example": "123"
            },
            "address": {
                "type": "string",
                "example": "Street 4"
            },
            "number": {
                "type": "string",
                "example": "123"
            },
            "cep": {
                "type": "string",
                "example": "78000-00"
            },
            "neighoarhood": {
                "type": "string",
                "example": "Bosque"
            },
            "countrie_id": {
                "type": "integer",
                "format": "int64",
                "example": "1"
            },
            "state_id": {
                "type": "integer",
                "format": "int64",
                "example": "1"
            },
            "citie_id": {
                "type": "integer",
                "format": "int64",
                "example": "1"
            },
            "status": {
                "type": "integer",
                "format": "int64",
                "example": "1"
            }
          },
          "xml": {
            "name": "Customer"
          }
        },
...
    }
}